refactor(core): remove unused Quartz dependencies and configuration
This commit removes the Quartz.AspNetCore package and its associated configuration from the `AppConfiguration.cs` file. The Quartz background job scheduling functionality is no longer required.

Additionally, it updates the `Nebx.BuildingBlocks.AspNetCore.csproj` file to reflect these removals and adjusts the version of `MediatR.Contracts` to `2.0.1`.

The `HttpFailureExtension.cs` file is updated to correctly reference `Failure` instead of `HttpFailure` in the `ToErrorResponse`