chore(deps): update FluentValidation.DependencyInjectionExtensions to 12.1.0
This commit updates the `FluentValidation.DependencyInjectionExtensions` package to version `12.1.0`.
It also updates the target framework to `net8.0` and the package version to `0.2.3` for the `Nebx.BuildingBlocks.AspNetCore` project and its tests.
Additionally, the `EventId` generation in `DomainEvent.cs` is changed from `Guid.CreateVersion7()` to `Guid.NewGuid()`.
A minor documentation correction is made in `SetupConfigurations.cs` to change "MediatR" to "mediator".